Extract and rename AlertsDate class
This is to be consistent with the new Alert class, which we will
create in the next commits.
