+++ lib/utils.py
@@ -2,37 +2,13 @@ import hashlib
 import pytz
 from datetime import datetime
 from pathlib import Path
-from pytz import timezone
+from lib.alert_date import AlertDate
 
 
 REPO = Path('.')
 SRC = REPO / 'src'
 DIST = REPO / 'dist'
 ROOT = DIST / 'alerts'
-
-
-class AlertsDate(object):
-    """Makes a datetime available in different formats"""
-
-    def __init__(self, _datetime):
-        self._datetime = _datetime
-        self._local_datetime = _datetime.astimezone(timezone('Europe/London'))
-
-    @property
-    def as_lang(self, lang='en-GB'):
-        return '{dt.day} {dt:%B} {dt:%Y} at {dt:%H}:{dt:%M}'.format(dt=self._local_datetime)
-
-    @property
-    def as_iso8601(self):
-        return self._local_datetime.isoformat()
-
-    @property
-    def as_datetime(self):
-        return self._datetime
-
-    @property
-    def as_local_datetime(self):
-        return self._local_datetime
 
 
 def file_fingerprint(path, root=DIST):
@@ -49,6 +25,6 @@ def is_current_alert(alert):
 
 
 def convert_dates(alert):
-    alert['sent'] = AlertsDate(alert['sent'])
-    alert['expires'] = AlertsDate(alert['expires'])
+    alert['sent'] = AlertDate(alert['sent'])
+    alert['expires'] = AlertDate(alert['expires'])
     return alert
