+++ tests/test_alert_date.py
@@ -1,0 +1,21 @@
+from datetime import datetime
+from lib.alert_date import AlertDate
+from pytz import timezone
+import pytz
+
+
+def test_AlertDate_properties():
+    sample_datetime = datetime(2021, 3, 21, 10, 30, tzinfo=pytz.utc)
+    alerts_date = AlertDate(sample_datetime)
+    assert alerts_date.as_lang == '21 March 2021 at 10:30'
+    assert alerts_date.as_iso8601 == '2021-03-21T10:30:00+00:00'
+    assert alerts_date.as_datetime == sample_datetime
+    assert alerts_date.as_local_datetime == sample_datetime  # sample date is outside of British Summer Time (BST)
+
+
+def test_AlertDate_properties_work_with_bst():
+    datetime_in_bst = datetime(2021, 4, 21, 10, 30, tzinfo=pytz.utc)
+    alerts_date = AlertDate(datetime_in_bst)
+    assert alerts_date.as_lang == '21 April 2021 at 11:30'
+    assert alerts_date.as_iso8601 == '2021-04-21T11:30:00+01:00'
+    assert alerts_date.as_datetime == datetime_in_bst.astimezone(timezone('Europe/London'))
