+++ tests/test_utils.py
@@ -3,26 +3,9 @@ import pytz
 
 from datetime import datetime
 from freezegun import freeze_time
-from lib.utils import AlertsDate, convert_dates, file_fingerprint, is_current_alert
+from lib.utils import convert_dates, file_fingerprint, is_current_alert
+from lib.alert_date import AlertDate
 from pathlib import Path
-from pytz import timezone
-
-
-def test_AlertsDate_properties():
-    sample_datetime = datetime(2021, 3, 21, 10, 30, tzinfo=pytz.utc)
-    alerts_date = AlertsDate(sample_datetime)
-    assert alerts_date.as_lang == '21 March 2021 at 10:30'
-    assert alerts_date.as_iso8601 == '2021-03-21T10:30:00+00:00'
-    assert alerts_date.as_datetime == sample_datetime
-    assert alerts_date.as_local_datetime == sample_datetime  # sample date is outside of British Summer Time (BST)
-
-
-def test_AlertsDate_properties_work_with_bst():
-    datetime_in_bst = datetime(2021, 4, 21, 10, 30, tzinfo=pytz.utc)
-    alerts_date = AlertsDate(datetime_in_bst)
-    assert alerts_date.as_lang == '21 April 2021 at 11:30'
-    assert alerts_date.as_iso8601 == '2021-04-21T11:30:00+01:00'
-    assert alerts_date.as_datetime == datetime_in_bst.astimezone(timezone('Europe/London'))
 
 
 def test_file_fingerprint_adds_hash_to_file_path():
@@ -42,7 +25,7 @@ def test_is_current_alert_checks_if_alert_is_current(expiry_date, is_current):
     assert is_current_alert(alert) == is_current
 
 
-def test_convert_dates_converts_alert_sent_and_expiry_dates_to_AlertsDate_class():
+def test_convert_dates_converts_alert_sent_and_expiry_dates_to_AlertDate_class():
     alert = {
         'message_type': 'alert',
         'expires': datetime(2021, 4, 21, 9, 30, tzinfo=pytz.utc),
@@ -50,5 +33,5 @@ def test_convert_dates_converts_alert_sent_and_expiry_dates_to_AlertsDate_class(
     }
     converted_alert = convert_dates(alert)
 
-    assert isinstance(converted_alert['sent'], AlertsDate)
-    assert isinstance(converted_alert['expires'], AlertsDate)
+    assert isinstance(converted_alert['sent'], AlertDate)
+    assert isinstance(converted_alert['expires'], AlertDate)
